started sorting the functions into individual files and afterwards importing them for ease
