+++ lectioscraper/lectioscraper.py
@@ -9,13 +9,50 @@ import time
 import pytz
 from typing import *
 import logging
+from getSchedule import get_schedule
+from getAbsence import get_absence
+from getAllHomework import get_all_homework
+from getAssignments import get_assignments
 
+class CustomFormatter(logging.Formatter):
+
+    grey = "\x1b[38;20m"
+    yellow = "\x1b[33;20m"
+    red = "\x1b[31;20m"
+    bold_red = "\x1b[31;1m"
+    green = "\x1b[32;20m"
+    reset = "\x1b[0m"
+    format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s (%(filename)s:%(lineno)d)"
+
+    FORMATS = {
+        logging.DEBUG: grey + format + reset,
+        logging.INFO: green + format + reset,
+        logging.WARNING: yellow + format + reset,
+        logging.ERROR: red + format + reset,
+        logging.CRITICAL: bold_red + format + reset
+    }
+
+    def format(self, record):
+        log_fmt = self.FORMATS.get(record.levelno)
+        formatter = logging.Formatter(log_fmt)
+        return formatter.format(record)
+
+logger = logging.getLogger("lectioscraper")
+logger.setLevel(logging.DEBUG)
+
+# create console handler with a higher log level
+ch = logging.StreamHandler()
+ch.setLevel(logging.DEBUG)
+
+ch.setFormatter(CustomFormatter())
+
+logger.addHandler(ch)
 
 class Lectio:
     def __init__(self, Username:str, Password:str, SchoolId:str):
         """
         Initializes the class with the username, password and school id.
-        Init is not made to be used directly, but is used when you create an instance of the class.
+        Init is not to be used directly, but is used when you create an instance of the class.
         
         :param Username: The username of the student.
         
@@ -64,10 +101,7 @@ class Lectio:
 
             self.Session = session
             initalized = True
-            
-            # print("Student id: " + self.studentId)
-            # print("School id: " + self.SchoolId)
-
+        
     def getSchedule(self, to_json:bool, print_to_console:bool=False):
         """
         getSchedule gets the schedule for the current week. Currently only works for the current week.
@@ -78,165 +112,9 @@ class Lectio:
 
         :return: if to_json is true, the schedule will be saved to a json file. If to_json is false, the schedule will be returned. If print_to_console is true, the schedule will be printed to the console.
         """
-
-        SCHEDULE_URL = "https://www.lectio.dk/lectio/" + self.SchoolId + "/SkemaNy.aspx?type=elev&elevid=" + self.studentId
-
-        schedule = self.Session.get(SCHEDULE_URL)
-
-        soup = BeautifulSoup(schedule.text, features="html.parser")
-        # print(soup.prettify())
-
-        scheduleContainer = soup.findAll('a', {"class": "s2bgbox"})
-
-        fullSchedule = []
-        Schedule = {}
-
-        # loop through the schedule and append the lessons to the fullSchedule list
-        if (scheduleContainer != None):
-            for schedule in scheduleContainer:
-                rows = schedule['data-additionalinfo'].split("\n")
-                timeStructure = re.compile(
-                    r'(\d+)/(\d+)-(\d+) (\d+):(\d+) til (\d+):(\d+)')
-                teamStructure = re.compile('Hold: ')
-                teacherStructure = re.compile('Lærer.*: ')
-                roomStructure = re.compile('Lokale.*: ')
-
-                #Getting the lesson id
-                # Get the lesson if normal
-                if "absid" in schedule['href']:
-                    lessonIdSplit1 = schedule['href'].split("absid=")
-                elif "ProeveholdId" in schedule['href']:
-                    lessonIdSplit1 = schedule['href'].split("ProeveholdId=")
-                else:
-                    logging.warning('Error')
-                    return False
-                
-                lessonIdSplit2 = lessonIdSplit1[1].split("&prevurl=")
-                lessonId = lessonIdSplit2[0]
-                
-                
-                #Check if there is a status
-                if rows[0] == "Aflyst!" or rows[0] == "Ændret!":
-                    #print("found a status: {}".format(rows[0]))
-
-                    status = rows[0]
-
-                    #Check if there is a title
-                    if timeStructure.match(rows[1]):
-                        #print("did not find a title")
-                        title = " "
-                    else:
-                        #print("found a title: {}".format(rows[1]))
-                        title = rows[1]
-
-                else:
-                    #print("did not find any status")
-                    status = " "
-
-                    #Check if there is a title
-                    if timeStructure.match(rows[0]):
-                        #print("did not find a title")
-                        title = " "
-                    else:
-                        #print("found a title: {}".format(rows[0]))
-                        title = rows[0]
-
-                time = list(filter(timeStructure.match, rows))
-                team = list(filter(teamStructure.match, rows))
-                teacher = list(filter(teacherStructure.match, rows))
-                room = list(filter(roomStructure.match, rows))
-
-                #If list is empty (There is no room or teacher) then make list empty
-                if len(time) == 0:
-                    time = " "
-                else:
-                    time = time[0]
-                
-                if len(team) == 0:
-                    team = " "
-                else:
-                    team = team[0].split(":")[1].strip()
-                
-                if len(teacher) == 0:
-                    teacher = " "
-                else:
-                    teacher = teacher[0].split(":")[1].strip()
-                
-                if len(room) == 0:
-                    room = " "
-                else:
-                    room = room[0].split(":")[1].strip()
-
-                #.split(":")[2]
-                time_split = time.split(" ")
-                
-                if status == "Aflyst!":
-                    Schedule['Status'] = "Aflyst!"
-                elif status == "Ændret!":
-                    Schedule['Status'] = "Aendret!"
-                else:
-                    Schedule['Status'] = "Normal!"
-                    
-                if title == " ":
-                    Schedule['Title'] = "Ingen titel"
-                else:
-                    Schedule['Title'] = title
-                    
-                Schedule['DateTime'] = time
-                Schedule['Date'] = time_split[0].replace("/", "-")
-                Schedule['StartTime'] = time_split[1]
-                Schedule['EndTime'] = time_split[3]
-                Schedule['Team'] = team
-                Schedule['Teacher'] = teacher
-                Schedule['Room'] = room
-                Schedule['Id'] = lessonId
-                
-
-
-                fullSchedule.append(Schedule)
-                Schedule = {}
-
-                
-                #DEBUG PURPOSES
-                """
-                print(time[0])
-                print(team[0])
-                print(teacher[0])
-                print(room[0])
-                
-                print("---------------------------")"""
-
-            
-            
-            if to_json == True:
-                # if a json file called schedule.json exists, it will be overwritten, if not it will be created
-                with open('schedule.json', 'w') as outfile:
-                    json_schedule = {}
-                    for lesson in fullSchedule:
-                        json_schedule[lesson['Id']] = lesson
-                    json.dump(json_schedule, outfile, indent=4)
-
-            elif to_json == False:
-                if print_to_console == True:
-                    logging.info("Schedule not saved, but it will be returned (print it to log it to the console)")
-                    return fullSchedule
-                else:
-                    return fullSchedule
-            else:
-                logging.warning("Please only provide boolean values in the to_json parameter")
-                return None
-        else:
-            logging.warning("No schedule found, please check your login details or see the readme")
-            boolean = False
-        return fullSchedule
-            
-            
         
-        
-        
-
-        
-        
+        return get_schedule(Session=self.Session, SchoolId=self.SchoolId, studentId=self.studentId, to_json=to_json)
+                            
     
     def getAbsence(self, written_assignments:bool, to_json:bool):
         """
@@ -247,48 +125,8 @@ class Lectio:
         
         :return: returns the absence for the student in different classes for the whole year.
         """
-
-        if type(to_json) is not bool:
-            logging.warning('to_json parameter must be a boolean')
-            return "to_json parameter must be a boolean"
-        if type(written_assignments) is not bool:
-            logging.warning('written_assignments parameter must be a boolean')
-            return "Must be a boolean"
-        ABSENCE_URL = "https://www.lectio.dk/lectio/{}/subnav/fravaerelev.aspx?elevid={}&prevurl=forside.aspx".format(self.SchoolId, self.studentId)
-        absence = self.Session.get(ABSENCE_URL)
-        soup = BeautifulSoup(absence.text, features="html.parser")
-        absence_table = soup.find("table", {"id": "s_m_Content_Content_SFTabStudentAbsenceDataTable"})
-        absence_rows = absence_table.findAll("tr")
-
-    
-    
+        return get_absence(Session=self.Session, SchoolId=self.SchoolId, studentId=self.studentId, written_assignments=written_assignments, to_json=to_json)
         
-        absence_end_result = []
-        
-        for i in absence_rows:
-            data = i.findAll("td")
-            absence = []
-            for j in data:
-                absence.append(j.text)
-            # print(absence)
-            if written_assignments is False:
-                if len(absence) == 9:
-                    absence_end_result.append({"team": absence[0], "opgjort": {"procent": absence[1], "moduler": absence[2]}, "for_the_year": {"procent": absence[3], "moduler": absence[4]}})
-            else:
-                if len(absence) == 9:
-
-                    absence_end_result.append({"team": absence[0], "opgjort": {"procent": absence[1], "moduler": absence[2]}, "for_the_year": {"procent": absence[3], "moduler": absence[4]}, "writing": {"opgjort": {"procent": absence[5], "elevtid": absence[6]}, "for_the_year_wrting": {"procent": absence[7], "elevtid": absence[8]}}})
-            
-        if to_json == True:
-            with open('absence.json', 'w') as outfile:
-                json.dump(absence_end_result, outfile, indent=4)
-        elif to_json == False:
-            logging.info("Absence not saved, but it will be returned (print it to log it to the console)")
-            return absence_end_result
-        else:
-            logging.warning("Please only provide boolean values in the to_json parameter")
-            return "to_json parameter must be a boolean"
-        return absence_end_result
     
     def getAllHomework(self, to_json:bool, print_to_console:bool):
         """
@@ -300,60 +138,7 @@ class Lectio:
         :return: returns all the homework in the 'lektier' either as a json file or as a dictionary.
     
         """
-        HOMEWORK_URL = "https://www.lectio.dk/lectio/{}/material_lektieoversigt.aspx?elevid={}".format(self.SchoolId, self.studentId)
-        homework = self.Session.get(HOMEWORK_URL)
-        soup = BeautifulSoup(homework.text, features="html.parser")
-        homework_table_div = soup.find("div", {"class": "ls-content"})
-        
-        table = soup.find("table", {"class": "ls-table-layout1"})
-        rows = table.findAll("tr")
-        homework_end_result = {}
-        
-        for i in rows:
-            date = i.find("th", {"class": "nowrap textTop"}).text
-            scheduleContainer = i.find('a', {"class": "s2bgbox"})
-            rows = scheduleContainer['data-additionalinfo'].split("\n")     
-            note = i.find_all('td', {"class": "textTop"})
-            for j in note:
-                if j.text != "":
-                    note = j.text
-                else:
-                    note = " "
-            # print(note)
-            
-            ls_homework = i.find("td", {"class": "ls-homework"})
-            # print(ls_homework.text)
-            time_split = rows[1].split(" ")
-            Homework = {
-                "modul": rows[0],
-                "date-and-time": rows[1],
-                "dato_for_modul": time_split[0].replace("/", "-"),
-                "startTid_for_modul": time_split[1],
-                "slutTid_for_modul": time_split[-1],
-                "hold": rows[2],
-                "teacher": rows[3],
-                "note": note,
-                "lektie": ls_homework.text
-            }
-            homework_end_result[rows[2].strip("Hold: ")] = Homework
-            
-        if to_json is False:
-            if print_to_console is True:
-                for i in homework_end_result:
-                    # print it in readable format
-                    print("Modul: {}".format(i['modul']))
-                    print("Date and time: {}".format(i['date-and-time']))
-                    print("Hold: {}".format(i['hold']))
-                    print("Teacher: {}".format(i['teacher']))
-                    print("Note: {}".format(i['note']))
-                    print("Lektie: {}".format(i['lektie']))
-                    print("---------------------------")
-            return homework_end_result
-            
-        else:
-            with open('homework.json', 'w') as outfile:
-                json.dump(homework_end_result, outfile, indent=4)
-            return homework_end_result
+        return get_all_homework(Session=self.Session, SchoolId=self.SchoolId, studentId=self.studentId, to_json=to_json, print_to_console=print_to_console)
         
     def getAssignments(self, to_json=False, team="alle hold", status="alle status", fravaer="", karakter=""):
         """
@@ -367,98 +152,7 @@ class Lectio:
 
         :return: returns all the assignments in the 'afleveringer' either as a json file or as a dictionary.
         """
-        ASSIGNMENT_URL = "https://www.lectio.dk/lectio/{}/OpgaverElev.aspx?elevid={}".format(self.SchoolId, self.studentId)
-        assignments = self.Session.get(ASSIGNMENT_URL)
-        soup = BeautifulSoup(assignments.text, features="html.parser")
-        assignment_table = soup.find("table", {"class": "ls-table-layout1 maxW textTop lf-grid"})
-        assignment_rows_with_headers = assignment_table.findAll("tr")
-        assignment_rows = assignment_rows_with_headers[1:]
-        assignment_end_result = [] 
-        for i in assignment_rows:
-            td = i.findAll("td")
-            assignment = []
-            for j in td:
-                assignment.append(j.text)
-            assignment_dict = {
-                "uge": assignment[0],
-                "hold": assignment[1],
-                "opgavetitel": assignment[2].strip("\n"),
-                "frist": assignment[3],
-                "tid": assignment[3][-5:],
-                "elevtid": assignment[4],
-                "status": assignment[5],
-                "fravaer": assignment[6],
-                "afventer": assignment[7],
-                "opgavenote": assignment[8],
-                "karakter": assignment[9],
-                "elevnote": assignment[10],
-            }
-            
-            assignment_end_result.append(assignment_dict)
-    
-        def filterAssignments(assignment_end_result, team, status, fravaer, karakter=""):
-            filtered_assignments = {}
-            for i in assignment_end_result:
-                if team != "alle hold" and status != "alle status" and fravaer != "" and karakter != "":
-                    if i["hold"] == team and i["status"] == status and i["fravaer"] == fravaer and i["karakter"] == karakter:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team != "alle hold" and status != "alle status" and fravaer != "" and karakter == "":
-                    if i["hold"] == team and i["status"] == status and i["fravaer"] == fravaer:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team != "alle hold" and status != "alle status" and fravaer == "" and karakter != "":
-                    if i["hold"] == team and i["status"] == status and i["karakter"] == karakter:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team != "alle hold" and status != "alle status" and fravaer == "" and karakter == "":
-                    if i["hold"] == team and i["status"] == status:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team != "alle hold" and status == "alle status" and fravaer != "" and karakter != "":
-                    if i["hold"] == team and i["fravaer"] == fravaer and i["karakter"] == karakter:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team != "alle hold" and status == "alle status" and fravaer != "" and karakter == "":
-                    if i["hold"] == team and i["fravaer"] == fravaer:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team != "alle hold" and status == "alle status" and fravaer == "" and karakter != "":
-                    if i["hold"] == team and i["karakter"] == karakter:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team != "alle hold" and status == "alle status" and fravaer == "" and karakter == "":
-                    if i["hold"] == team:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team == "alle hold" and status != "alle status" and fravaer != "" and karakter != "":
-                    if i["status"] == status and i["fravaer"] == fravaer and i["karakter"] == karakter:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team == "alle hold" and status != "alle status" and fravaer != "" and karakter == "":
-                    if i["status"] == status and i["fravaer"] == fravaer:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team == "alle hold" and status != "alle status" and fravaer == "" and karakter != "":
-                    if i["status"] == status and i["karakter"] == karakter:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team == "alle hold" and status != "alle status" and fravaer == "" and karakter == "":
-                    if i["status"] == status:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team == "alle hold" and status == "alle status" and fravaer != "" and karakter != "":
-                    if i["fravaer"] == fravaer and i["karakter"] == karakter:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team == "alle hold" and status == "alle status" and fravaer != "" and karakter == "":
-                    if i["fravaer"] == fravaer:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team == "alle hold" and status == "alle status" and fravaer == "" and karakter != "":
-                    if i["karakter"] == karakter:
-                        filtered_assignments[i["opgavetitel"]] = i
-                elif team == "alle hold" and status == "alle status" and fravaer == "" and karakter == "":
-                    filtered_assignments[i["opgavetitel"]] = i
-            # print(len(filtered_assignments))
-            if len(filtered_assignments) == 0:
-                logging.info('No assignments found with the given filters, please check your filters')
-                return "No assignments found with the given filters, please check your filters"
-            else:
-                if to_json is False:
-                    return filtered_assignments
-                else:
-                    with open('assignments.json', 'w') as outfile:
-                        json.dump(filtered_assignments, outfile, indent=4)
-                    return filtered_assignments
-        
-        return filterAssignments(assignment_end_result, team, status, fravaer, karakter)
+        return get_assignments(Session=self.Session, SchoolId=self.SchoolId, studentId=self.studentId, to_json=to_json, team=team, status=status, fravaer=fravaer, karakter=karakter)
     
     def getTodaysSchedule(self, to_json=False):
         """
@@ -596,3 +290,5 @@ class Lectio:
             
     
 
+client = Lectio("emil763x", "xf96G?YteTYqEA#A", "59")
+client.getAssignments(to_json=True, team="1g/3 MaB")
