+++ requirements.txt
@@ -1,3 +1,42 @@
-sphinx
-sphinx_autodoc_typehints
-sphinx_rtd_theme
+alabaster==0.7.12
+Babel==2.10.3
+beautifulsoup4==4.11.1
+bleach==5.0.1
+certifi==2022.6.15
+charset-normalizer==2.1.1
+commonmark==0.9.1
+docutils==0.19
+idna==3.3
+imagesize==1.4.1
+importlib-metadata==4.12.0
+Jinja2==3.1.2
+keyring==23.8.2
+lxml==4.9.1
+MarkupSafe==2.1.1
+packaging==21.3
+pkginfo==1.8.3
+Pygments==2.13.0
+pyparsing==3.0.9
+pytz==2022.2.1
+readme-renderer==37.0
+requests==2.28.1
+requests-toolbelt==0.9.1
+rfc3986==2.0.0
+rich==12.5.1
+six==1.16.0
+snowballstemmer==2.2.0
+soupsieve==2.3.2.post1
+Sphinx==5.1.1
+sphinx-autodoc-typehints==1.19.2
+sphinx-rtd-theme==1.0.0
+sphinxcontrib-applehelp==1.0.2
+sphinxcontrib-devhelp==1.0.2
+sphinxcontrib-htmlhelp==2.0.0
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==1.0.3
+sphinxcontrib-serializinghtml==1.1.5
+twine==4.0.1
+typing_extensions==4.3.0
+urllib3==1.26.12
+webencodings==0.5.1
+zipp==3.8.1
