--- schedule.json
@@ -1,242 +1,0 @@
-{
-    "54243396500": {
-        "Status": "Normal!",
-        "Title": "Du, ich und wir",
-        "DateTime": "15/8-2022 08:20 til 09:50",
-        "Date": "15-8-2022",
-        "StartTime": "08:20",
-        "EndTime": "09:50",
-        "Team": "1g TyB/gf-1",
-        "Teacher": "Anna Marbjerg (AM)",
-        "Room": "A13",
-        "Id": "54243396500"
-    },
-    "54243396610": {
-        "Status": "Normal!",
-        "Title": "1: INTRODUKTION",
-        "DateTime": "15/8-2022 10:05 til 11:35",
-        "Date": "15-8-2022",
-        "StartTime": "10:05",
-        "EndTime": "11:35",
-        "Team": "1g/3 latin",
-        "Teacher": "Catherine Lizzie Vigh (CV)",
-        "Room": "B25",
-        "Id": "54243396610"
-    },
-    "54243396757": {
-        "Status": "Aflyst!",
-        "Title": "Ingen titel",
-        "DateTime": "15/8-2022 12:05 til 13:35",
-        "Date": "15-8-2022",
-        "StartTime": "12:05",
-        "EndTime": "13:35",
-        "Team": "1g/3 EnB",
-        "Teacher": "Kasper Timm Tornbjerg (KT)",
-        "Room": "B25",
-        "Id": "54243396757"
-    },
-    "54243397026": {
-        "Status": "Normal!",
-        "Title": "Modul 1",
-        "DateTime": "16/8-2022 08:20 til 09:50",
-        "Date": "16-8-2022",
-        "StartTime": "08:20",
-        "EndTime": "09:50",
-        "Team": "1g/3 DAA",
-        "Teacher": "Louise Marie Ralfs (LM)",
-        "Room": "B25",
-        "Id": "54243397026"
-    },
-    "54243397160": {
-        "Status": "Normal!",
-        "Title": "Ingen titel",
-        "DateTime": "16/8-2022 10:05 til 11:35",
-        "Date": "16-8-2022",
-        "StartTime": "10:05",
-        "EndTime": "11:35",
-        "Team": "1g/3 saC",
-        "Teacher": "Kirsten Kj\u00e6r S\u00f8rensen (KS)",
-        "Room": "B25",
-        "Id": "54243397160"
-    },
-    "54877862579": {
-        "Status": "Normal!",
-        "Title": "Frokostskak",
-        "DateTime": "16/8-2022 11:35 til 12:05",
-        "Date": "16-8-2022",
-        "StartTime": "11:35",
-        "EndTime": "12:05",
-        "Team": "Alle 1. G. elever, Alle 2. G. elever, Alle 3. G. elever",
-        "Teacher": "Anders Biel (AB)",
-        "Room": "Fs",
-        "Id": "54877862579"
-    },
-    "54807117602": {
-        "Status": "Normal!",
-        "Title": "Fotografering",
-        "DateTime": "16/8-2022 10:35 til 11:05",
-        "Date": "16-8-2022",
-        "StartTime": "10:35",
-        "EndTime": "11:05",
-        "Team": "1g/3 au",
-        "Teacher": "Kirsten Kj\u00e6r S\u00f8rensen (KS)",
-        "Room": "YK",
-        "Id": "54807117602"
-    },
-    "54243397291": {
-        "Status": "Normal!",
-        "Title": "Introduktion til faget",
-        "DateTime": "16/8-2022 12:05 til 13:35",
-        "Date": "16-8-2022",
-        "StartTime": "12:05",
-        "EndTime": "13:35",
-        "Team": "1g/3 ap",
-        "Teacher": "Anna Marbjerg (AM)",
-        "Room": "B25",
-        "Id": "54243397291"
-    },
-    "54243397524": {
-        "Status": "Normal!",
-        "Title": "Ingen titel",
-        "DateTime": "17/8-2022 08:20 til 09:50",
-        "Date": "17-8-2022",
-        "StartTime": "08:20",
-        "EndTime": "09:50",
-        "Team": "1g/3 idC",
-        "Teacher": "Milan Lorenzen (ML)",
-        "Room": "ID",
-        "Id": "54243397524"
-    },
-    "54243397660": {
-        "Status": "Normal!",
-        "Title": "Intro 2 LM",
-        "DateTime": "17/8-2022 10:05 til 11:35",
-        "Date": "17-8-2022",
-        "StartTime": "10:05",
-        "EndTime": "11:35",
-        "Team": "1g/3 nv",
-        "Teacher": "KW, LM, PA",
-        "Room": "C5",
-        "Id": "54243397660"
-    },
-    "54813192490": {
-        "Status": "Aendret!",
-        "Title": "RG SPORT+ M\u00d8DE",
-        "DateTime": "17/8-2022 11:35 til 11:55",
-        "Date": "17-8-2022",
-        "StartTime": "11:35",
-        "EndTime": "11:55",
-        "Team": "Alle 1. G. elever, Alle 2. G. elever, Alle 3. G. elever",
-        "Teacher": "JH, KW",
-        "Room": "BK",
-        "Id": "54813192490"
-    },
-    "54243397936": {
-        "Status": "Normal!",
-        "Title": "Ingen titel",
-        "DateTime": "17/8-2022 13:50 til 15:20",
-        "Date": "17-8-2022",
-        "StartTime": "13:50",
-        "EndTime": "15:20",
-        "Team": "1g meC/2",
-        "Teacher": "Cecilie Blach Renneberg (Bl)",
-        "Room": "B10",
-        "Id": "54243397936"
-    },
-    "54876522699": {
-        "Status": "Normal!",
-        "Title": "Strikkeklub",
-        "DateTime": "17/8-2022 11:36 til 12:04",
-        "Date": "17-8-2022",
-        "StartTime": "11:36",
-        "EndTime": "12:04",
-        "Team": "Alle 1. G. elever, Alle 2. G. elever, Alle 3. G. elever",
-        "Teacher": "Le, LM",
-        "Room": "ST.C",
-        "Id": "54876522699"
-    },
-    "54243397802": {
-        "Status": "Normal!",
-        "Title": "Ingen titel",
-        "DateTime": "17/8-2022 12:05 til 13:35",
-        "Date": "17-8-2022",
-        "StartTime": "12:05",
-        "EndTime": "13:35",
-        "Team": "1g/3 MaB",
-        "Teacher": "Michael Falk Hansen (FH)",
-        "Room": "B18",
-        "Id": "54243397802"
-    },
-    "54243398050": {
-        "Status": "Normal!",
-        "Title": "Ingen titel",
-        "DateTime": "18/8-2022 10:05 til 11:35",
-        "Date": "18-8-2022",
-        "StartTime": "10:05",
-        "EndTime": "11:35",
-        "Team": "1g/3 MaB",
-        "Teacher": "Michael Falk Hansen (FH)",
-        "Room": "B18",
-        "Id": "54243398050"
-    },
-    "54243398186": {
-        "Status": "Normal!",
-        "Title": "Modul 2",
-        "DateTime": "18/8-2022 12:05 til 13:35",
-        "Date": "18-8-2022",
-        "StartTime": "12:05",
-        "EndTime": "13:35",
-        "Team": "1g/3 DAA",
-        "Teacher": "Louise Marie Ralfs (LM)",
-        "Room": "B25",
-        "Id": "54243398186"
-    },
-    "54243398320": {
-        "Status": "Normal!",
-        "Title": "Intro 3 LM",
-        "DateTime": "18/8-2022 13:50 til 15:20",
-        "Date": "18-8-2022",
-        "StartTime": "13:50",
-        "EndTime": "15:20",
-        "Team": "1g/3 nv",
-        "Teacher": "KW, LM, PA",
-        "Room": "C14",
-        "Id": "54243398320"
-    },
-    "54243398573": {
-        "Status": "Normal!",
-        "Title": "Bio 1 LM",
-        "DateTime": "19/8-2022 10:05 til 11:35",
-        "Date": "19-8-2022",
-        "StartTime": "10:05",
-        "EndTime": "11:35",
-        "Team": "1g/3 nv",
-        "Teacher": "KW, LM, PA",
-        "Room": "C5",
-        "Id": "54243398573"
-    },
-    "54243398704": {
-        "Status": "Normal!",
-        "Title": "2: IMPERIUM ROMANUM",
-        "DateTime": "19/8-2022 12:05 til 13:35",
-        "Date": "19-8-2022",
-        "StartTime": "12:05",
-        "EndTime": "13:35",
-        "Team": "1g/3 latin",
-        "Teacher": "Catherine Lizzie Vigh (CV)",
-        "Room": "B25",
-        "Id": "54243398704"
-    },
-    "54243398836": {
-        "Status": "Normal!",
-        "Title": "Nabosprog",
-        "DateTime": "19/8-2022 13:50 til 15:20",
-        "Date": "19-8-2022",
-        "StartTime": "13:50",
-        "EndTime": "15:20",
-        "Team": "1g/3 ap",
-        "Teacher": "Anna Marbjerg (AM)",
-        "Room": "B25",
-        "Id": "54243398836"
-    }
-}
